requirements.txt
numpy
pandas 
scipy
xarray
#tiled[client]
netcdf4
jupyterlab
ipywidgets
ipympl
nodejs
matplotlib
seaborn
pint
periodictable
flask
flask-jwt-extended >=4.0
bokeh
pillow
scikit-image
pyserial
requests
flask_cors
#pyFAI
#certif-pyspec==1.1.12

